requirements.txt
appdirs==1.4.4
astroid==2.6.6
attrs==21.2.0
black==21.7b0
boto3==1.18.14
botocore==1.21.14
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
idna==3.2
iniconfig==1.1.1
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
pluggy==1.0.0
py==1.10.0
pylint==2.9.6
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
quickchart.io==0.2.0
regex==2021.8.3
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
slack-bolt==1.7.0
slack-sdk==3.8.0
toml==0.10.2
tomli==1.2.0
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
